Chem heater: Make borgs use the same UI as humans
Instead of using magic to heat (cool) the beaker outside of the machine,
the borg will now insert the beaker into it like a normal human would,
and manipulate the heater with the standard (HTML) interface. To avoid
bullshit, only the borg that inserted the beaker is able to eject it. To
avoid even more bullshit, moving the borg out of range will also eject
the beaker. Obviously, "ejecting" the beaker doesn't place it on turf as
a standard object; it stays inside the borg's module at all times.