Borgs: Keep tools in the module hud when equipping
Removing the tool from the module hud would alter its ordering and make
things confusing. Instead, we'll just always show all tools and make the
ones that are already equipped non-interactable and grayed out.

This involves a bunch of updates all across the code because BYOND
doesn't support displaying an atom in multiple places in the UI, so we
have to send icon change notifications from all objs that are
(potentially) part of a borg module.